refactor: record_link_changes を concern に切り出し UnitsController の行数超過を解消
